Fix format string error on x86_32
Ref. Componolit/gart#60

+++ liblog/tests/log_read_test.cpp
@@ -42,8 +42,8 @@ TEST(liblog, __android_log_write__android_logger_list_read) {
 
   struct timespec ts;
   clock_gettime(CLOCK_MONOTONIC, &ts);
-  std::string buf = android::base::StringPrintf("pid=%u ts=%" PRId64 ".%09ld", pid,
-                                                ts.tv_sec, ts.tv_nsec);
+  std::string buf = android::base::StringPrintf("pid=%u ts=%ld.%09ld", pid,
+                                                (long)ts.tv_sec, ts.tv_nsec);
   static const char tag[] =
       "liblog.__android_log_write__android_logger_list_read";
   static const char prio = ANDROID_LOG_DEBUG;
